add additional attribute 'key' to Net::SAML2::SP
Do not rely on the fact that the file given as the 'cert' parameter
also contains the private key.

Add an additional attribute 'key' to the Net::SAML2::SP object that holds the
filename of the private key for the certificate file given with the parameter
'cert'.

This change allows using the standard certificate & private key files
in lots of unix distributions (e.g. Debian, where certificate files are
in /etc/ssl/certs/ and private key files are in /etc/ssl/private/)